:bug: JpaRepository 제네릭 파라미터 Long 으로 변경 #93

+++ vsplay/src/main/java/com/buck/vsplay/domain/vstopic/repository/VsTopicRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VsTopicRepository extends JpaRepository<VsTopic, Integer> {
+public interface VsTopicRepository extends JpaRepository<VsTopic, Long> {
 }
